+++ src/test/java/lotto/model/LottoCollectionTest.java
@@ -1,16 +1,14 @@
-package lotto;
+package lotto.model;
 
-import lotto.model.LottoCollection;
-import lotto.model.Winning;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
 import java.util.Arrays;
 import java.util.List;
 
+import static lotto.model.Lotto.toLottoNumber;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoCollectionTest {
@@ -20,10 +18,10 @@ class LottoCollectionTest {
 
     @BeforeEach
     void setup() {
-        List<Integer> lotto1 = List.of(3, 5, 10, 12, 16, 33);
-        List<Integer> lotto2 = List.of(4, 6, 10, 11, 13, 18);
-        List<Integer> lotto3 = List.of(7, 8, 14, 18, 19, 35);
-        List<List<Integer>> lottos = List.of(lotto1, lotto2, lotto3);
+        Lotto lotto1 = toLottoNumber(List.of(3, 5, 10, 12, 16, 33));
+        Lotto lotto2 = toLottoNumber(List.of(4, 6, 10, 11, 13, 18));
+        Lotto lotto3 = toLottoNumber(List.of(7, 8, 14, 18, 19, 35));
+        List<Lotto> lottos = List.of(lotto1, lotto2, lotto3);
 
         collection = new LottoCollection(lottos);
         numbers = Arrays.asList("1", "2", "3", "4", "5", "6");
@@ -45,7 +43,7 @@ class LottoCollectionTest {
     @CsvSource(value = {"3,1", "4,1", "5,0", "6,0"})
     @DisplayName("일치하는 번호개수를 카운팅한다")
     void test3(int element, int expected) {
-        Winning winning = new Winning(collection, Arrays.asList("3", "5", "10", "11", "13", "18"));
-        assertThat(winning.getMatchingCount(integer -> integer == element)).isEqualTo(expected);
+        WinningNumber winningNumber = new WinningNumber(collection, Arrays.asList("3", "5", "10", "11", "13", "18"));
+        assertThat(winningNumber.getMatchingCount(integer -> integer == element)).isEqualTo(expected);
     }
 }
