+++ src/test/java/lotto/model/WinningNumberTest.java
@@ -1,7 +1,5 @@
-package lotto;
+package lotto.model;
 
-import lotto.model.LottoCollection;
-import lotto.model.Winning;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -13,9 +11,9 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class WinningTest {
+class WinningNumberTest {
 
-    private final List<List<Integer>> list = new ArrayList<>();
+    private final List<Lotto> list = new ArrayList<>();
     private final LottoCollection collection = new LottoCollection(list);
 
     @ParameterizedTest
@@ -23,7 +21,7 @@ class WinningTest {
     @DisplayName("로또번호는 1과 45 사이값이어야만 한다")
     void winningNumberBoundTest(String input) {
         assertThatThrownBy(() -> {
-            new Winning(collection, Arrays.asList(input, "2", "3", "4", "5", "6"));
+            new WinningNumber(collection, Arrays.asList(input, "2", "3", "4", "5", "6"));
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -31,7 +29,7 @@ class WinningTest {
     @DisplayName("로또번호는 6개를 입력해야한다")
     void winningNumberQuantityTest() {
         assertThatThrownBy(() -> {
-            new Winning(collection, Arrays.asList("2", "3", "4", "5", "6"));
+            new WinningNumber(collection, Arrays.asList("2", "3", "4", "5", "6"));
         }).isInstanceOf(IllegalArgumentException.class);
     }
 }
